Rename factory files
