+++ src/backend/tests/expungement_analyzer/test_expunger.py
@@ -2,7 +2,7 @@ import unittest
 
 from datetime import datetime
 from expungeservice.expungement_analyzer.expunger import Expunger
-from tests.factories.crawler import CrawlerFactory
+from tests.factories.crawler_factory import CrawlerFactory
 from tests.fixtures.case_details import CaseDetails
 from tests.fixtures.john_doe import JohnDoe
